form validation

+++ server/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const cors = require('cors')
+const cors = require('cors');
 const MongoClient = require('mongodb').MongoClient;
 
 const app = express();
@@ -8,7 +8,7 @@ const port = process.env.PORT || 5000;
 const url = 'mongodb+srv://nodejsconnect:nodejsconnect@cluster0.qqhi9.mongodb.net/assignment?retryWrites=true&w=majority';
 let database, collection;
 
-app.use(bodyParser.json())
+app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true}));
 app.use(cors())
 
@@ -16,6 +16,7 @@ MongoClient.connect(url, { useNewUrlParser: true }, (error, client) => {
     if(error) {
         throw error;
     }
+    
     database = client.db("assignment");
     collection = database.collection("users");
     console.log("Connected to database!");
@@ -23,30 +24,44 @@ MongoClient.connect(url, { useNewUrlParser: true }, (error, client) => {
 
 app.get('/api/show', async (req, res) => {
     const usersData = await collection.find().toArray();
-    console.log(`Total users ${usersData.length}`, usersData)
+    console.log(`Total users ${usersData.length}`);
     res.send({ usersData });
 });
 
 app.post('/api/add', async (req, res) => {
     const { emailId } = req.body;
+    const count =  await collection.find({ emailId }).count();
 
-    await collection.insertOne({ emailId })
+    if(count > 0) {
+        res.send(`Cannot add, user already exists`);
+    } 
+    else{
+        users = await collection.insertOne({ emailId });
+        const usersData = await collection.find().toArray();
 
-    console.log(req.body, "added sucessfully");
-    res.send(collection.find().toArray());
+        res.send({message:`User added successfully, Total users: ${usersData.length}`, users});
+        console.log({ emailId }, `Added successfully, Total users: ${usersData.length}`);
+    }
 });
 
 app.delete('/api/delete', async (req, res) => {
-    console.log(mash)
     const { emailId } = req.body;
-    console.log({ emailId })
-    await collection.deleteOne({ emailId })
+    const count =  await collection.find({ emailId }).count();
 
-    console.log(req.body, "deleted sucessfully");
-    res.send(collection.find().toArray());
+    if(count > 0) {
+        users = await collection.deleteOne({ emailId });
+        const usersData = await collection.find().toArray();
+
+        res.send({message:`User deleted successfully, Total users: ${usersData.length}`, users});
+        console.log({ emailId }, `Deletd successfully, Total users: ${usersData.length}`);
+    } 
+    else {
+        res.send(`Cannot delete, user does not exist`);
+    }
+    
 });
 
 app.listen(port, () => {
-    console.log(`Example app listening at port ${port}`)
+    console.log(`Example app listening at port ${port}`);
   })
